refactor: createBridgeClient
