+++ examples/child-process/src/lib/vite/runtime/bun.js
@@ -1,60 +1,11 @@
-import assert from "node:assert";
-import { ESModulesEvaluator, ModuleRunner } from "vite/module-runner";
+import { createBridgeClient } from "../bridge-client.js";
 
 async function main() {
   // @ts-ignore
-  const { bridgeUrl, root } = JSON.parse(process.argv[2]);
-
-  /**
-   * @param {string} method
-   * @param  {...any} args
-   * @returns {Promise<any>}
-   */
-  async function bridgeRpc(method, ...args) {
-    const response = await fetch(bridgeUrl + "/rpc", {
-      method: "POST",
-      body: JSON.stringify({ method, args }),
-    });
-    assert(response.ok);
-    const result = response.json();
-    return result;
-  }
-
-  const runner = new ModuleRunner(
-    {
-      root,
-      sourcemapInterceptor: "prepareStackTrace",
-      transport: {
-        fetchModule: (...args) => bridgeRpc("fetchModule", ...args),
-      },
-      hmr: false,
-    },
-    new ESModulesEvaluator(),
-  );
-
-  /**
-   * @param {Request} request
-   * @returns {Promise<Response>}
-   */
-  async function handler(request) {
-    try {
-      const headers = request.headers;
-      // @ts-ignore
-      const meta = JSON.parse(headers.get("x-vite-meta"));
-      headers.delete("x-vite-meta");
-      const mod = await runner.import(meta.entry);
-      return mod.default(new Request(meta.url, { ...request, headers }));
-    } catch (e) {
-      console.error(e);
-      const message =
-        "[bun runner error]\n" +
-        (e instanceof Error ? `${e.stack ?? e.message}` : "");
-      return new Response(message, { status: 500 });
-    }
-  }
-
-  const server = Bun.serve({ port: 0, fetch: handler });
-  await bridgeRpc("register", `http://localhost:${server.port}`);
+  const options = JSON.parse(process.argv[2]);
+  const bridgeClient = createBridgeClient(options);
+  const server = Bun.serve({ port: 0, fetch: bridgeClient.handler });
+  await bridgeClient.rpc("register", `http://localhost:${server.port}`);
 }
 
 main();
