+++ examples/child-process/src/lib/vite/runtime/node.js
@@ -1,62 +1,14 @@
 import assert from "node:assert";
 import http from "node:http";
 import { webToNodeHandler } from "@hiogawa/utils-node";
-import { ESModulesEvaluator, ModuleRunner } from "vite/module-runner";
+import { createBridgeClient } from "../bridge-client.js";
 
 async function main() {
   // @ts-ignore
-  const { bridgeUrl, root } = JSON.parse(process.argv[2]);
+  const options = JSON.parse(process.argv[2]);
+  const bridgeClient = createBridgeClient(options);
 
-  /**
-   *
-   * @param {string} method
-   * @param  {...any} args
-   * @returns
-   */
-  async function bridgeRpc(method, ...args) {
-    const response = await fetch(bridgeUrl + "/rpc", {
-      method: "POST",
-      body: JSON.stringify({ method, args }),
-    });
-    assert(response.ok);
-    const result = response.json();
-    return result;
-  }
-
-  const runner = new ModuleRunner(
-    {
-      root,
-      sourcemapInterceptor: "prepareStackTrace",
-      transport: {
-        fetchModule: (...args) => bridgeRpc("fetchModule", ...args),
-      },
-      hmr: false,
-    },
-    new ESModulesEvaluator(),
-  );
-
-  /**
-   * @param {Request} request
-   * @returns {Promise<Response>}
-   */
-  async function handler(request) {
-    try {
-      const headers = request.headers;
-      // @ts-ignore
-      const meta = JSON.parse(headers.get("x-vite-meta"));
-      headers.delete("x-vite-meta");
-      const mod = await runner.import(meta.entry);
-      return mod.default(new Request(meta.url, { ...request, headers }));
-    } catch (e) {
-      console.error(e);
-      const message =
-        "[node runner error]\n" +
-        (e instanceof Error ? `${e.stack ?? e.message}` : "");
-      return new Response(message, { status: 500 });
-    }
-  }
-
-  const listener = webToNodeHandler(handler);
+  const listener = webToNodeHandler(bridgeClient.handler);
 
   const server = http.createServer((req, res) => {
     listener(req, res, (e) => console.error(e));
@@ -65,7 +17,7 @@ async function main() {
   server.listen(async () => {
     const address = server.address();
     assert(address && typeof address !== "string");
-    await bridgeRpc("register", `http://localhost:${address.port}`);
+    await bridgeClient.rpc("register", `http://localhost:${address.port}`);
   });
 }
