+++ examples/child-process/src/lib/vite/types.ts
@@ -1,0 +1,4 @@
+export interface BridgeClientOptions {
+  bridgeUrl: string;
+  root: string;
+}
